+++ src/index.js
@@ -18,6 +18,20 @@ import I18nPath from './path'
 
 import type { PathValue } from './path'
 
+const numberFormatKeys = [
+  'style',
+  'currency',
+  'currencyDisplay',
+  'useGrouping',
+  'minimumIntegerDigits',
+  'minimumFractionDigits',
+  'maximumFractionDigits',
+  'minimumSignificantDigits',
+  'maximumSignificantDigits',
+  'localeMatcher',
+  'formatMatcher'
+]
+
 export default class VueI18n {
   static install: () => void
   static version: string
@@ -39,6 +53,14 @@ export default class VueI18n {
   _dataListeners: Array<any>
 
   constructor (options: I18nOptions = {}) {
+    // Auto install if it is not done yet and `window` has `Vue`.
+    // To allow users to avoid auto-installation in some cases,
+    // this code should be placed here. See #290
+    /* istanbul ignore if */
+    if (!Vue && typeof window !== 'undefined' && window.Vue) {
+      install(window.Vue)
+    }
+
     const locale: Locale = options.locale || 'en-US'
     const fallbackLocale: Locale = options.fallbackLocale || 'en-US'
     const messages: LocaleMessages = options.messages || {}
@@ -147,7 +169,7 @@ export default class VueI18n {
   _getDateTimeFormats (): DateTimeFormats { return this._vm.dateTimeFormats }
   _getNumberFormats (): NumberFormats { return this._vm.numberFormats }
 
-  _warnDefault (locale: Locale, key: Path, result: ?any, vm: ?any): ?string {
+  _warnDefault (locale: Locale, key: Path, result: ?any, vm: ?any, values: any): ?string {
     if (!isNull(result)) { return result }
     if (this.missing) {
       const missingResult = this.missing.call(null, locale, key, vm)
@@ -180,7 +202,7 @@ export default class VueI18n {
     if (!message) { return null }
 
     const pathRet: PathValue = this._path.getPathValue(message, key)
-    if (Array.isArray(pathRet)) { return pathRet }
+    if (Array.isArray(pathRet) || isPlainObject(pathRet)) { return pathRet }
 
     let ret: mixed
     if (isNull(pathRet)) {
@@ -213,7 +235,7 @@ export default class VueI18n {
       ret = this._link(locale, message, ret, host, interpolateMode, values)
     }
 
-    return !values ? ret : this._render(ret, interpolateMode, values)
+    return this._render(ret, interpolateMode, values)
   }
 
   _link (
@@ -258,7 +280,10 @@ export default class VueI18n {
           linkPlaceholder, host, interpolateMode, values
         )
       }
-      translated = this._warnDefault(locale, linkPlaceholder, translated, host)
+      translated = this._warnDefault(
+        locale, linkPlaceholder, translated, host,
+        Array.isArray(values) ? values : [values]
+      )
 
       // Replace the link with the translated
       ret = !translated ? ret : ret.replace(link, translated)
@@ -316,7 +341,7 @@ export default class VueI18n {
       if (!this._root) { throw Error('unexpected error') }
       return this._root.t(key, ...values)
     } else {
-      return this._warnDefault(locale, key, ret, host)
+      return this._warnDefault(locale, key, ret, host, values)
     }
   }
 
@@ -334,7 +359,7 @@ export default class VueI18n {
       if (!this._root) { throw Error('unexpected error') }
       return this._root.i(key, locale, values)
     } else {
-      return this._warnDefault(locale, key, ret, host)
+      return this._warnDefault(locale, key, ret, host, [values])
     }
   }
 
@@ -386,7 +411,7 @@ export default class VueI18n {
   }
 
   mergeLocaleMessage (locale: Locale, message: LocaleMessageObject): void {
-    this._vm.messages[locale] = Vue.util.extend(this._vm.messages[locale] || {}, message)
+    this._vm.$set(this._vm.messages, locale, Vue.util.extend(this._vm.messages[locale] || {}, message))
   }
 
   getDateTimeFormat (locale: Locale): DateTimeFormat {
@@ -398,7 +423,7 @@ export default class VueI18n {
   }
 
   mergeDateTimeFormat (locale: Locale, format: DateTimeFormat): void {
-    this._vm.dateTimeFormats[locale] = Vue.util.extend(this._vm.dateTimeFormats[locale] || {}, format)
+    this._vm.$set(this._vm.dateTimeFormats, locale, Vue.util.extend(this._vm.dateTimeFormats[locale] || {}, format))
   }
 
   _localizeDateTime (
@@ -436,7 +461,7 @@ export default class VueI18n {
   _d (value: number | Date, locale: Locale, key: ?string): DateTimeFormatResult {
     /* istanbul ignore if */
     if (process.env.NODE_ENV !== 'production' && !VueI18n.availabilities.dateTimeFormat) {
-      warn('Cannot format a Date value due to not support Intl.DateTimeFormat.')
+      warn('Cannot format a Date value due to not supported Intl.DateTimeFormat.')
       return ''
     }
 
@@ -494,7 +519,7 @@ export default class VueI18n {
   }
 
   mergeNumberFormat (locale: Locale, format: NumberFormat): void {
-    this._vm.numberFormats[locale] = Vue.util.extend(this._vm.numberFormats[locale] || {}, format)
+    this._vm.$set(this._vm.numberFormats, locale, Vue.util.extend(this._vm.numberFormats[locale] || {}, format))
   }
 
   _localizeNumber (
@@ -502,7 +527,8 @@ export default class VueI18n {
     locale: Locale,
     fallback: Locale,
     numberFormats: NumberFormats,
-    key: string
+    key: string,
+    options: ?NumberFormatOptions
   ): ?NumberFormatResult {
     let _locale: Locale = locale
     let formats: NumberFormat = numberFormats[_locale]
@@ -520,35 +546,43 @@ export default class VueI18n {
       return null
     } else {
       const format: ?NumberFormatOptions = formats[key]
-      const id = `${_locale}__${key}`
-      let formatter = this._numberFormatters[id]
-      if (!formatter) {
-        formatter = this._numberFormatters[id] = new Intl.NumberFormat(_locale, format)
+
+      let formatter
+      if (options) {
+        // If options specified - create one time number formatter
+        formatter = new Intl.NumberFormat(_locale, Object.assign({}, format, options))
+      } else {
+        const id = `${_locale}__${key}`
+        formatter = this._numberFormatters[id]
+        if (!formatter) {
+          formatter = this._numberFormatters[id] = new Intl.NumberFormat(_locale, format)
+        }
       }
       return formatter.format(value)
     }
   }
 
-  _n (value: number, locale: Locale, key: ?string): NumberFormatResult {
+  _n (value: number, locale: Locale, key: ?string, options: ?NumberFormatOptions): NumberFormatResult {
     /* istanbul ignore if */
     if (process.env.NODE_ENV !== 'production' && !VueI18n.availabilities.numberFormat) {
-      warn('Cannot format a Date value due to not support Intl.NumberFormat.')
+      warn('Cannot format a Number value due to not supported Intl.NumberFormat.')
       return ''
     }
 
     if (!key) {
-      return new Intl.NumberFormat(locale).format(value)
+      const nf = !options ? new Intl.NumberFormat(locale) : new Intl.NumberFormat(locale, options)
+      return nf.format(value)
     }
 
     const ret: ?NumberFormatResult =
-      this._localizeNumber(value, locale, this.fallbackLocale, this._getNumberFormats(), key)
+      this._localizeNumber(value, locale, this.fallbackLocale, this._getNumberFormats(), key, options)
     if (this._isFallbackRoot(ret)) {
       if (process.env.NODE_ENV !== 'production') {
         warn(`Fall back to number localization of root: key '${key}' .`)
       }
       /* istanbul ignore if */
       if (!this._root) { throw Error('unexpected error') }
-      return this._root.n(value, key, locale)
+      return this._root.n(value, Object.assign({}, { key, locale }, options))
     } else {
       return ret || ''
     }
@@ -557,6 +591,7 @@ export default class VueI18n {
   n (value: number, ...args: any): NumberFormatResult {
     let locale: Locale = this.locale
     let key: ?string = null
+    let options: ?NumberFormatOptions = null
 
     if (args.length === 1) {
       if (typeof args[0] === 'string') {
@@ -568,6 +603,14 @@ export default class VueI18n {
         if (args[0].key) {
           key = args[0].key
         }
+
+        // Filter out number format options only
+        options = Object.keys(args[0]).reduce((acc, key) => {
+          if (numberFormatKeys.includes(key)) {
+            return Object.assign({}, acc, { [key]: args[0][key] })
+          }
+          return acc
+        }, null)
       }
     } else if (args.length === 2) {
       if (typeof args[0] === 'string') {
@@ -578,7 +621,7 @@ export default class VueI18n {
       }
     }
 
-    return this._n(value, locale, key)
+    return this._n(value, locale, key, options)
   }
 }
 
@@ -588,8 +631,3 @@ VueI18n.availabilities = {
 }
 VueI18n.install = install
 VueI18n.version = '__VERSION__'
-
-/* istanbul ignore if */
-if (typeof window !== 'undefined' && window.Vue) {
-  window.Vue.use(VueI18n)
-}