+++ test/unit/format_custom.test.js
@@ -17,6 +17,22 @@ describe('custom formatter', () => {
       })
       i18n.t('message.hello', 'ja', { name: 'joe' })
     })
+
+    it('should interpolate messages without values', done => {
+      class CustomFormatter {
+        interpolate (message, values) {
+          assert(values === null)
+          done()
+        }
+      }
+      const formatter = new CustomFormatter()
+      const i18n = new VueI18n({
+        locale: 'en',
+        messages,
+        formatter
+      })
+      i18n.t('message.hello')
+    })
   })
 
   describe('via vue instance calling', () => {
@@ -35,6 +51,49 @@ describe('custom formatter', () => {
         })
       })
       vm.$t('message.hello', [1, 2, 3])
+    })
+  })
+
+  describe('via vue instance calling (mounted)', () => {
+    let el
+
+    beforeEach(done => {
+      el = document.createElement('div')
+      done()
+    })
+
+    it('should be inherited by components', done => {
+      new Vue({
+        i18n: new VueI18n({
+          locale: 'en',
+          formatter: {
+            interpolate: (message, values) => {
+              assert.deepEqual({ name: 'user' }, values)
+              done()
+              return ['pass']
+            }
+          }
+        }),
+        components: {
+          'child-1': {
+            render (h) {
+              return h('div', {}, [
+                h('p', {}, [this.$t('message', { name: 'user' })])
+              ])
+            },
+            i18n: {
+              messages: {
+                en: { message: 'hello {name}' }
+              }
+            }
+          }
+        },
+        render (h) {
+          return h('div', {}, [
+            h('child-1')
+          ])
+        }
+      }).$mount(el)
     })
   })